feature: Embedded API server to interact with operator views (#4)
This change adds a custom Kubernetes extension API server that can be used to interact with the view
resources maintained internally by dcontroller operators. This lets a standard kubectl client to
connect to the dcontroller manager and get, list, modify or watch view resources.

This PR introduces a significant change in the way operators maintain views:
- each operator gets its own API group <operator-name>.view.dcontroller.io
- the API server delegates API requests to the operator caches via the a client multiplexer
- the client multiplexer arbitrates API requests across the controller-runtime clients of each
  operator's cache based on the requested API group

Currently the API server is exposed over insecure unauthenticated HTTP for simplicity.